+++ src/Business/StandardItem.java
@@ -2,7 +2,7 @@ package Business;
 
 public class StandardItem extends Item {
 
-    public StandardItem(int id, String name, int power, int durability, String classe) {
+    public StandardItem(long id, String name, int power, int durability, String classe) {
         super(id, name, power, durability, classe);
     }
