+++ src/Business/SuperItem.java
@@ -2,7 +2,7 @@ package Business;
 
 public class SuperItem extends Item{
 
-    public SuperItem(int id, String name, int power, int durability, String classe) {
+    public SuperItem(long id, String name, int power, int durability, String classe) {
         super(id, name, power, durability, classe);
     }
